355: avoid some problems due to having Xpect in the product
Since Xpect is also installed in the product, all files that contain the
string "XPECT", case sensitive, are handled by Xpect in special ways
(file extension does not matter). This can lead to problems during
PluginUI tests and therefore this PR is changing some tests to not use
the problematic substring in project names.