#355: Clean up Xpect tests in bundle org.eclipse.n4js.n4jsx.xpect.ui.tests (#356)
* 355: repair & clean up tests in org.eclipse.n4js.n4jsx.xpect.ui.tests

* 355: remove some debug code

* 355: handle IEObjectDescriptionWithError in predicate of Xpect scope

* 355: remove some changes only used for experimentation/debugging

* 355: fix & improve another test in another bundle

* 355: avoid some problems due to having Xpect in the product

Since Xpect is also installed in the product, all files that contain the
string "XPECT", case sensitive, are handled by Xpect in special ways
(file extension does not matter). This can lead to problems during
PluginUI tests and therefore this PR is changing some tests to not use
the problematic substring in project names.