GH-658: Hover supports formatting of JSDoc including some markdown
As a user I need JSDoc hovers to be formatted so that I can better read
the documentation.
HTML is supported out of the box, but I also need minimal support for
tags and markdown.

The hovers should be aware of JSDoc tags. Some markdown features should
be supported to beatify typical comments:

be aware of paragraphs (emptyline)
unordered lists via "-"
Tags should be added below the description, using definition list as in
JDT.