Refactored builds that are triggered for external projects
There was a race between the AutoBuildJob and the jobs that are scheduled
to reconcile external projects after a local workspace project was
opened or closed.

The logic was changed such that all the external changes are processed
in a single operation rather than multiple jobs. The pending changes
are also observable from the outside.

Signed-off-by: Sebastian Zarnekow <sebastian.zarnekow@gmail.com>