Fixed a bunch of potential races in the builder.
The handling of project references was changes. Previously we
relied completely on the dynamic references to be available when
the build starts. This assumption does not hold, since these references
are set in an asynchronous job. There was code, that was supposed to
mitigate the issue, but that was flawed, too, since it cause unnecessary
cache invalidation.

The builder now obtains the deltas for all referenced projects of the
currently built project. It checks, whether the referenced project is still
present and if yes, it performs as before. If the referenced project was closed
or removed from the workspace, the builder takes the pending deltas into account.

It will also ensure, that the build is triggered correctly for the pending external
changes such that it sees an up-to-date index.

Signed-off-by: Sebastian Zarnekow <sebastian.zarnekow@gmail.com>