GH-1409: fixed a bug when name of repo name appears multiple times in path
The JSDoc2Adoc feature requires the repository relative location of
files. The resolution searched the repository and the project name in
the absolute path of a file. This may fail for two reasons:

1) the repository name may be different from the clonded folder
2) any of the two names may appear multiple times in the path, e.g. in
"oomph/n4js/git/n4js/project/src/org/eclipse/n4js/file"
the repository name "n4js" appears three times.

This it is not sufficient to look for simple names. Instead, only the
paths are used now. This may lead to problems in case of linked files,
though. So we need to keep track of that.