GH-1497: As a user I would like the transpiler to optimize trivial cases of reflection usage (#1499)
* transpiler optimizes trivial usage of the reflection APIs

* add tests

* proper handling for built-in types + improve a test

* transpiler optimizes access to name of functions (incl. constructors)

This commit moves the ES6(!) getter 'name' from n4js-runtime-es2015 into
the built-in type Function, to allow access to it from the transpiler.
This is justified by the fact that we have done this several times
already in the past, e.g. for built-in type Promise.

* add tests

* typo

* roll back commits 8fd3546d and e04b6d7f

Rationale: we cannot perform this optimization because it is based on
the assumption that functions will have the same name at runtime as they
have in the N4JS source code, but minifiers, etc. will break this
assumption.