GH-1561: Reduce memory footprint of LSP headless builder (#1579)
This change ensures that resources are not loaded redundantly in the LSP builder case but
resolved from the dependent project's associated resource set instead.

When a project is built initially, all its resources are loaded into a resource set including their AST. After the project build is done, the resource set is cleared.

Downstream projects in the dependency graph will attempt to use the resources they depend. These are loaded from module descriptions into the resource set of their owner project, not into the resource set of the currently built project. Subsequent projects may reuse these loaded resources, that don't have an AST, without the need to load them again. Without the AST, the resources are very slim and can be kept in memory. This is a temporary solution to improve the headless build.

Also the caching of validation issues was improved in this change set.

see #1561